backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
pydantic
psycopg2-binary
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
fastapi-users